[2235] Reduce the amount of code in DiagramRenderer
Bug: https://github.com/eclipse-sirius/sirius-web/issues/2235
